Dashboard: rename site-launch-celebration-modal folder to site-launch-modal
The folder now holds the shared SiteLaunchModal (both the pre-launch and
celebration variants), its styles, tests, and stories, so the celebration-only
name no longer fits. The SiteLaunchCelebrationModal export is unchanged; only
the directory and import paths are renamed.
